feat(calculateAnswers): Refactor WriteProgression to render key names based on keyNamesText in lib/data.
