+++ app/lib/data/keyNamesText.js
@@ -1,0 +1,10 @@
+const keyNames = [
+  "C major",
+  "E minor",
+  "Db major",
+  "C# minor",
+  "D# major",
+  "Gb minor",
+];
+
+export default keyNames;
